+++ solver/gcr.c
@@ -170,11 +170,11 @@ static void init_gcr(const int _M, const int _V){
     xi = calloc(M, sizeof(spinor *));
 #if (defined SSE || defined SSE2 || defined SSE3)
     _a = calloc((M+2)*M, sizeof(complex));
-    a[0] = (complex *)(((unsigned int)(_a)+ALIGN_BASE)&~ALIGN_BASE); 
+    a[0] = (complex *)(((unsigned long int)(_a)+ALIGN_BASE)&~ALIGN_BASE); 
     _chi = calloc(M*Vo+1, sizeof(spinor));
-    chi[0] = (spinor *)(((unsigned int)(_chi)+ALIGN_BASE)&~ALIGN_BASE);
+    chi[0] = (spinor *)(((unsigned long int)(_chi)+ALIGN_BASE)&~ALIGN_BASE);
     _xi = calloc(M*Vo+1, sizeof(spinor));
-    xi[0] = (spinor *)(((unsigned int)(_xi)+ALIGN_BASE)&~ALIGN_BASE);
+    xi[0] = (spinor *)(((unsigned long int)(_xi)+ALIGN_BASE)&~ALIGN_BASE);
 #else
     _a = calloc((M+1)*M, sizeof(complex));
     a[0] = _a;
